Overview: Filelist displays the name of the card now.
Detailed: Filelist displayed the path list used to load the file before, now it displays the name. When it is double clicked the index of the file list is used to access the file path stored in a list variable within the window class. Same loading functionality, except now a name is displayed instead of a filepath.